BaseFragment 생성, BindingAdapter object 삭제 및 extension 추가
