+++ wkdehdlr/CoinHelper/app/src/main/java/kr/schoolsharing/coinhelper/base/BaseFragment.kt
@@ -1,0 +1,25 @@
+package kr.schoolsharing.coinhelper.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+
+abstract class BaseFragment<B : ViewDataBinding>(private val layoutId: Int) : Fragment() {
+    protected lateinit var binding: B
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate(inflater, layoutId, container, false)
+        binding.lifecycleOwner = this
+
+        return binding.root
+    }
+
+}
