+++ wkdehdlr/CoinHelper/app/src/main/java/kr/schoolsharing/coinhelper/tasks/PagerAdapter.kt
@@ -2,9 +2,9 @@ package kr.schoolsharing.coinhelper.tasks
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentStatePagerAdapter
+import androidx.fragment.app.FragmentPagerAdapter
 
-class PagerAdapter(fm: FragmentManager) : FragmentStatePagerAdapter(fm) {
+class PagerAdapter(fm: FragmentManager) : FragmentPagerAdapter(fm) {
 
     private val marketList = listOf("KRW", "BTC", "ETH", "USDT")
