+++ wkdehdlr/CoinHelper/app/src/main/java/kr/schoolsharing/coinhelper/tasks/UpbitFragment.kt
@@ -1,33 +1,17 @@
 package kr.schoolsharing.coinhelper.tasks
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import kr.schoolsharing.coinhelper.R
+import kr.schoolsharing.coinhelper.base.BaseFragment
 import kr.schoolsharing.coinhelper.data.Repository
 import kr.schoolsharing.coinhelper.databinding.FragmentCoinrecyclerBinding
 
-class UpbitFragment : Fragment() {
+class UpbitFragment : BaseFragment<FragmentCoinrecyclerBinding>(R.layout.fragment_coinrecycler) {
 
     private lateinit var upbitViewModel: UpbitViewModel
-    private lateinit var binding: FragmentCoinrecyclerBinding
 
     private val rVAdapter = UpbitRVAdapter()
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        binding = DataBindingUtil.inflate(
-            LayoutInflater.from(context),
-            R.layout.fragment_coinrecycler,
-            container,
-            false
-        )
-        return binding.root
-    }
-    
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
