+++ wkdehdlr/CoinHelper/app/src/main/java/kr/schoolsharing/coinhelper/util/BindingAdapters.kt
@@ -1,0 +1,14 @@
+package kr.schoolsharing.coinhelper.util
+
+import androidx.databinding.BindingAdapter
+import androidx.databinding.ObservableField
+import androidx.recyclerview.widget.RecyclerView
+import kr.schoolsharing.coinhelper.model.UpbitItem
+import kr.schoolsharing.coinhelper.tasks.UpbitRVAdapter
+
+@BindingAdapter("items")
+fun RecyclerView.replaceAll(items: ObservableField<List<UpbitItem>>) {
+    (this.adapter as? UpbitRVAdapter)?.let {
+        it.setTickerList(items.get()!!)
+    }
+}
