+++ wkdehdlr/CoinHelper/app/src/main/res/layout/fragment_coinrecycler.xml
@@ -5,6 +5,8 @@
 
     <data>
 
+        <import type="android.view.View" />
+
         <variable
             name="viewModel"
             type="kr.schoolsharing.coinhelper.tasks.UpbitViewModel" />
